auth delete route

+++ app/Http/Controllers/AnnouncementController.php
@@ -77,7 +77,12 @@ class AnnouncementController extends Controller
      */
     public function destroy(Announcement $announcement)
     {
-        $announcement->delete(); 
+        if($announcement->user_id == Auth::user()->id ){
+            $announcement->delete(); 
+        }
+        else {
+            abort(404);
+        }
         return redirect()->back()->with('delete', 'Annuncio eliminato');
     }
 }
